[experimental] Initial implementation of `cbuild-run.yml` handling (#1096)
