+++ tools/projmgr/include/ProjMgrYamlEmitter.h
@@ -7,7 +7,15 @@
 #ifndef PROJMGRYAMLEMITTER_H
 #define PROJMGRYAMLEMITTER_H
 
+#include "ProjMgrRunDebug.h"
 #include "ProjMgrWorker.h"
+
+/**
+ * Forward declarations
+*/
+namespace YAML {
+  class Node;
+}
 
 /**
   * @brief projmgr emitter yaml class
@@ -17,7 +25,7 @@ public:
   /**
    * @brief class constructor
   */
-  ProjMgrYamlEmitter(void);
+  ProjMgrYamlEmitter(ProjMgrParser* parser, ProjMgrWorker* worker);
 
   /**
    * @brief class destructor
@@ -25,67 +33,86 @@ public:
   ~ProjMgrYamlEmitter(void);
 
   /**
+   * @brief set check schema
+   * @param bool checkSchema
+  */
+  void SetCheckSchema(bool checkSchema);
+
+  /**
+   * @brief set output directory
+   * @param string outputDir
+  */
+  void SetOutputDir(const std::string& outputDir);
+
+  /**
    * @brief generate cbuild-idx.yml file
-   * @param parser reference
-   * @param worker reference
    * @param contexts vector with pointers to contexts
-   * @param outputDir directory
    * @param failed contexts
    * @param executes nodes at solution level
-   * @param boolean check schema of generated file
    * @return true if executed successfully
   */
-  static bool GenerateCbuildIndex(ProjMgrParser& parser, ProjMgrWorker& worker,
-    const std::vector<ContextItem*>& contexts, const std::string& outputDir,
-    const std::set<std::string>& failedContexts,
-    const std::map<std::string, ExecutesItem>& executes, bool checkSchema);
+  bool GenerateCbuildIndex(const std::vector<ContextItem*>& contexts,
+    const std::set<std::string>& failedContexts, const std::map<std::string, ExecutesItem>& executes);
 
   /**
    * @brief generate cbuild-gen-idx.yml file
-   * @param parser reference
    * @param contexts vector with pointers to sibling contexts
    * @param type project type
    * @param outdir output directory
    * @param gendir generated files directory
-   * @param boolean check schema of generated file
    * @return true if executed successfully
   */
-  static bool GenerateCbuildGenIndex(ProjMgrParser& parser, const std::vector<ContextItem*> siblings,
-    const std::string& type, const std::string& outdir, const std::string& gendir, bool checkSchema);
+  bool GenerateCbuildGenIndex(const std::vector<ContextItem*> siblings,
+    const std::string& type, const std::string& outdir, const std::string& gendir);
 
   /**
    * @brief generate cbuild.yml or cbuild-gen.yml file
    * @param context pointer to the context
-   * @param boolean check schema of generated file
    * @param reference to generator identifier
    * @param reference to generator pack
+   * @param ignoreRteFileMissing if true, ignore missing rte files
    * @return true if executed successfully
   */
-  static bool GenerateCbuild(ContextItem* context, bool checkSchema,
-    const std::string& generatorId = std::string(),
-    const std::string& generatorPack = std::string(),
-    bool ignoreRteFileMissing = false);
+  bool GenerateCbuild(ContextItem* context, const std::string& generatorId = std::string(),
+    const std::string& generatorPack = std::string(), bool ignoreRteFileMissing = false);
 
   /**
    * @brief generate cbuild set file
    * @param contexts list of selected contexts
    * @param selectedCompiler name of the selected compiler
    * @param cbuildSetFile path to *.cbuild-set file to be generated
-   * @param boolean check schema of generated file
+   * @param ignoreRteFileMissing if true, ignore missing rte files
    * @return true if executed successfully
   */
-  static bool GenerateCbuildSet(const std::vector<std::string> selectedContexts,
-    const std::string& selectedCompiler, const std::string& cbuildSetFile, bool checkSchema, bool ignoreRteFileMissing = false);
+  bool GenerateCbuildSet(const std::vector<std::string> selectedContexts,
+    const std::string& selectedCompiler, const std::string& cbuildSetFile, bool ignoreRteFileMissing = false);
 
   /**
    * @brief generate cbuild pack file
    * @param contexts vector with pointers to contexts
    * @param keepExistingPackContent if true, all entries from existing cbuild-pack should be preserved
    * @param cbuildPackFrozen if true, reject updates to cbuild pack file
-   * @param boolean check schema of generated file
    * @return true if executed successfully
   */
-  static bool GenerateCbuildPack(ProjMgrParser& parser, const std::vector<ContextItem*> contexts, bool keepExistingPackContent, bool cbuildPackFrozen, bool checkSchema);
+  bool GenerateCbuildPack(const std::vector<ContextItem*> contexts, bool keepExistingPackContent, bool cbuildPackFrozen);
+
+  /**
+   * @brief generate cbuild run file
+   * @param reference to struct with debug/run management info
+   * @return true if executed successfully
+  */
+  bool GenerateCbuildRun(const RunDebugType& debugRun);
+
+protected:
+  ProjMgrParser* m_parser = nullptr;
+  ProjMgrWorker* m_worker = nullptr;
+  std::string m_outputDir;
+  bool m_checkSchema;
+
+  bool WriteFile(YAML::Node& rootNode, const std::string& filename, const std::string& context = std::string(), bool allowUpdate = true);
+  bool CompareFile(const std::string& filename, const YAML::Node& rootNode);
+  bool CompareNodes(const YAML::Node& lhs, const YAML::Node& rhs);
+  bool NeedRebuild(const std::string& filename, const YAML::Node& rootNode);
 };
 
 #endif  // PROJMGRYAMLEMITTER_H
