+++ tools/projmgr/include/ProjMgrYamlParser.h
@@ -13,6 +13,8 @@
 /**
   * @brief YAML key definitions
 */
+static constexpr const char* YAML_ACCESS = "access";
+static constexpr const char* YAML_ALGORITHM = "algorithm";
 static constexpr const char* YAML_APIS = "apis";
 static constexpr const char* YAML_API = "api";
 static constexpr const char* YAML_ADDPATH = "add-path";
@@ -40,6 +42,7 @@ static constexpr const char* YAML_CBUILD = "cbuild";
 static constexpr const char* YAML_CBUILD_GENS = "cbuild-gens";
 static constexpr const char* YAML_CBUILD_GEN = "cbuild-gen";
 static constexpr const char* YAML_CBUILD_PACK = "cbuild-pack";
+static constexpr const char* YAML_CBUILD_RUN = "cbuild-run";
 static constexpr const char* YAML_CBUILD_SET = "cbuild-set";
 static constexpr const char* YAML_CDEFAULT = "cdefault";
 static constexpr const char* YAML_CLAYERS = "clayers";
@@ -114,6 +117,7 @@ static constexpr const char* YAML_LICENSES = "licenses";
 static constexpr const char* YAML_LICENSE_AGREEMENT = "license-agreement";
 static constexpr const char* YAML_LINKER = "linker";
 static constexpr const char* YAML_MAP = "map";
+static constexpr const char* YAML_MEMORY = "memory";
 static constexpr const char* YAML_MESSAGES = "messages";
 static constexpr const char* YAML_MISC = "misc";
 static constexpr const char* YAML_MISC_ASM = "ASM";
@@ -144,10 +148,13 @@ static constexpr const char* YAML_PACKS_MISSING = "packs-missing";
 static constexpr const char* YAML_PACKS_UNUSED = "packs-unused";
 static constexpr const char* YAML_PATH = "path";
 static constexpr const char* YAML_PROCESSOR = "processor";
+static constexpr const char* YAML_PROGRAMMING = "programming";
 static constexpr const char* YAML_PROJECT = "project";
 static constexpr const char* YAML_PROJECTS = "projects";
 static constexpr const char* YAML_PROJECT_TYPE = "project-type";
 static constexpr const char* YAML_PROVIDES = "provides";
+static constexpr const char* YAML_RAM_SIZE = "ram-size";
+static constexpr const char* YAML_RAM_START = "ram-start";
 static constexpr const char* YAML_REBUILD = "rebuild";
 static constexpr const char* YAML_REGIONS = "regions";
 static constexpr const char* YAML_RESOLVED_PACK = "resolved-pack";
@@ -165,8 +172,11 @@ static constexpr const char* YAML_SETUP = "setup";
 static constexpr const char* YAML_SET = "set";
 static constexpr const char* YAML_SETTINGS = "settings";
 static constexpr const char* YAML_SELECT_COMPILER = "select-compiler";
+static constexpr const char* YAML_SIZE = "size";
+static constexpr const char* YAML_START = "start";
 static constexpr const char* YAML_STATUS = "status";
 static constexpr const char* YAML_SWITCH = "switch";
+static constexpr const char* YAML_SYSTEM_DESCRIPTIONS = "system-descriptions";
 static constexpr const char* YAML_TARGET_CONFIGURATIONS = "target-configurations";
 static constexpr const char* YAML_TARGETTYPE = "target-type";
 static constexpr const char* YAML_TARGETTYPES = "target-types";
