+++ tools/projmgr/schemas/cbuild-run.schema.json
@@ -1,0 +1,15 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$id": "https://raw.githubusercontent.com/Open-CMSIS-Pack/devtools/schemas/projmgr/2.6.0/tools/projmgr/schemas/cbuild-run.schema.json",
+  "title": "CMSIS cbuild-run",
+  "description": "contains run and debug information for a target",
+  "version": "2.6.0",
+  "type": "object",
+  "properties": {
+    "cbuild-run": {
+      "$ref": "./common.schema.json#/definitions/RunDebugDescType"
+    }
+  },
+  "additionalProperties": false,
+  "required": [ "cbuild-run" ]
+}
