+++ tools/projmgr/src/ProjMgrYamlParser.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2021 Arm Limited. All rights reserved.
+ * Copyright (c) 2020-2024 Arm Limited. All rights reserved.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -913,6 +913,20 @@ bool ProjMgrYamlParser::ParseTargetType(const YAML::Node& parent, const string& 
       }
     }
   }
+
+  if (parent[YAML_MEMORY].IsDefined()) {
+    const YAML::Node& memoryNode = parent[YAML_MEMORY];
+    for (const auto& memoryEntry : memoryNode) {
+      MemoryItem memoryItem;
+      ParseString(memoryEntry, YAML_NAME, memoryItem.name);
+      ParseString(memoryEntry, YAML_ACCESS, memoryItem.access);
+      ParseString(memoryEntry, YAML_START, memoryItem.start);
+      ParseString(memoryEntry, YAML_SIZE, memoryItem.size);
+      ParseString(memoryEntry, YAML_ALGORITHM, memoryItem.algorithm);
+      targetType.memory.push_back(memoryItem);
+    }
+  }
+
   return ParseBuildType(parent, file, targetType.build);
 }
 
@@ -1055,6 +1069,7 @@ const set<string> targetTypeKeys = {
   YAML_TYPE,
   YAML_DEVICE,
   YAML_BOARD,
+  YAML_MEMORY,
   YAML_PROCESSOR,
   YAML_COMPILER,
   YAML_OPTIMIZE,
