+++ tools/projmgr/test/data/TestRunDebug/run-debug.cproject.yml
@@ -1,0 +1,7 @@
+# yaml-language-server: $schema=https://raw.githubusercontent.com/Open-CMSIS-Pack/devtools/main/tools/projmgr/schemas/cproject.schema.json
+
+project:
+
+  components:
+    - component: Startup
+    - component: CORE
