+++ tools/projmgr/test/src/ProjMgrUnitTests.cpp
@@ -6572,3 +6572,18 @@ TEST_F(ProjMgrUnitTests, InvalidContextSet) {
   auto errStr = streamRedirect.GetErrorString();
   EXPECT_NE(string::npos, errStr.find("unknown selected context(s):\n  unknown1.debug+target\n  unknown2.release+target"));
 }
+
+TEST_F(ProjMgrUnitTests, TestRunDebug) {
+  char* argv[6];
+  const string& csolution = testinput_folder + "/TestRunDebug/run-debug.csolution.yml";
+  argv[1] = (char*)"convert";
+  argv[2] = (char*)csolution.c_str();
+  argv[3] = (char*)"--context-set";
+  argv[4] = (char*)"-o";
+  argv[5] = (char*)testoutput_folder.c_str();
+  //argv[6] = (char*)"-n";
+  EXPECT_EQ(0, RunProjMgr(6, argv, m_envp));
+
+  ProjMgrTestEnv::CompareFile(testoutput_folder + "/TestHW.cbuild-run.yml",
+    testinput_folder + "/TestRunDebug/ref/TestHW.cbuild-run.yml");
+}
